worker/requirements.txt
pika==1.1.0
redis==3.3.7
pyyaml==5.1.2